[REMANIEMENT] Utilise le lien reçu dans le corps de la réponse lors de l'appel à l'API utilisateur pour se connecter
