+++ mon-aide-cyber-ui/test/fournisseurs/reducteurUtilisateurAuthentifie.spec.tsx
@@ -1,15 +1,29 @@
-import { describe, expect, it } from 'vitest';
-
-import { Navigate, Outlet } from 'react-router-dom';
+import { describe, expect } from 'vitest';
 import {
   initialiseReducteurUtilisateurAuthentifie,
   reducteurUtilisateurAuthentifie,
   utilisateurCharge,
   utilisateurNonAuthentifie,
-} from '../../src/fournisseurs/ContexteAuthentification.tsx';
+} from '../../src/fournisseurs/reducteurUtilisateurAuthentifie.tsx';
+import { unUtilisateur } from '../constructeurs/constructeurUtilisateur.ts';
+import { Navigate, Outlet } from 'react-router-dom';
 
-describe('Fournisseur contexte authentification', () => {
-  it("l'utilisateur n'est pas authentifié", () => {
+describe('Réducteur utilisateur authentifié', () => {
+  it("charge l'utilisateur", () => {
+    const utilisateur = unUtilisateur().construis();
+    const etatUtilisateurAuthentifie = reducteurUtilisateurAuthentifie(
+      initialiseReducteurUtilisateurAuthentifie(),
+      utilisateurCharge(utilisateur),
+    );
+
+    expect(etatUtilisateurAuthentifie).toStrictEqual({
+      element: <Outlet />,
+      enAttenteDeChargement: false,
+      utilisateur,
+    });
+  });
+
+  it("redirige l'utilisateur", () => {
     const etatUtilisateurAuthentifie = reducteurUtilisateurAuthentifie(
       initialiseReducteurUtilisateurAuthentifie(),
       utilisateurNonAuthentifie(),
@@ -18,21 +32,6 @@ describe('Fournisseur contexte authentification', () => {
     expect(etatUtilisateurAuthentifie).toStrictEqual({
       element: <Navigate to="/" />,
       enAttenteDeChargement: false,
-    });
-  });
-
-  it('est authentifié', () => {
-    const etatUtilisateurAuthentifie = reducteurUtilisateurAuthentifie(
-      initialiseReducteurUtilisateurAuthentifie(),
-      utilisateurCharge({ nomPrenom: 'Jean Dupont' }),
-    );
-
-    expect(etatUtilisateurAuthentifie).toStrictEqual({
-      element: <Outlet />,
-      enAttenteDeChargement: false,
-      utilisateur: {
-        nomPrenom: 'Jean Dupont',
-      },
     });
   });
 });
